[feat/chat] feat: add ChatMessageSendDTO(for Time)

+++ src/main/java/cos/peerna/domain/ConnectedUser.java
@@ -23,11 +23,11 @@ public class ConnectedUser {
 
     @Id
     private Long id;
-    private Long roomId;
+    private Integer roomId;
     private boolean proceedAgree;
     private LocalDateTime lastConnectedAt;
 
-    public ConnectedUser(Long id, Long roomId) {
+    public ConnectedUser(Long id, Integer roomId) {
         this.id = id;
         this.roomId = roomId;
         this.proceedAgree = false;
